dev_requirements.txt
alabaster==0.7.16
astroid==3.1.0
Babel==2.14.0
black==24.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
dill==0.3.8
docutils==0.21.1
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
Pygments==2.17.2
pylint==3.1.0
pyserial==3.5
pytest==8.1.1
pytest-cov==5.0.0
requests==2.31.0
smbus2==0.4.3
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
tomlkit==0.12.4
urllib3==2.2.1
